添加关卡0-1的json文件

+++ ArknightSimulator/Operation/EnemyMovement.cs
@@ -9,7 +9,7 @@ namespace ArknightSimulator.Operation
     public class EnemyMovement
     {
         public EnemyTemplate Enemy { get; set; }
-        public int EntryTime { get; set; }
+        public float EntryTime { get; set; }
         public List<Point> MovingPoints { get; set; }
